+++ src/main/java/io/contek/invoker/deribit/api/common/_IndexPrice.java
@@ -1,9 +1,8 @@
 package io.contek.invoker.deribit.api.common;
 
-import java.math.BigDecimal;
 
 public class _IndexPrice {
 
   public String estimated_delivery_price;
-  public BigDecimal index_price;
+  public Double index_price;
 }
