+++ src/main/java/io/contek/invoker/deribit/api/common/_Instrument.java
@@ -1,7 +1,6 @@
 package io.contek.invoker.deribit.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _Instrument {
@@ -11,7 +10,7 @@ public class _Instrument {
   public String instrument_name;
   public int instrument_id;
   public String block_trade_commission;
-  public BigDecimal contract_size;
+  public Double contract_size;
   public Long creation_timestamp;
   public Long expiration_timestamp;
   public boolean is_active;
@@ -19,9 +18,9 @@ public class _Instrument {
   public Long leverage;
   public String maker_commission;
   public String taker_commission;
-  public BigDecimal min_trade_amount;
+  public Double min_trade_amount;
   public String option_type;
   public String settlement_period;
   public String strike;
-  public BigDecimal tick_size;
+  public Double tick_size;
 }
