+++ src/main/java/io/contek/invoker/deribit/api/common/_SubAccountSummary.java
@@ -1,6 +1,5 @@
 package io.contek.invoker.deribit.api.common;
 
-import java.math.BigDecimal;
 import java.util.Map;
 
 public class _SubAccountSummary {
@@ -14,9 +13,9 @@ public class _SubAccountSummary {
 
     public static class _PortfolioPosition {
         public String currency;
-        public BigDecimal equity;
-        public BigDecimal spot_reserve;
-        public BigDecimal initial_margin;
-        public BigDecimal maintenance_margin;
+        public Double equity;
+        public Double spot_reserve;
+        public Double initial_margin;
+        public Double maintenance_margin;
     }
 }
